Major rearrangement of files
