+++ aea/connections/scaffold/connection.yaml
@@ -1,0 +1,9 @@
+name: scaffold
+authors: Fetch.AI Limited
+version: 0.1.0
+license: Apache 2.0
+url: ""
+class_name: MyScaffoldConnection
+supported_protocols: []
+config:
+  foo: bar
