+++ aea/protocols/oef/serialization.py
@@ -24,8 +24,8 @@ import copy
 import json
 import pickle
 
-from aea.protocols.base.message import Message
-from aea.protocols.base.serialization import Serializer
+from aea.protocols.base import Message
+from aea.protocols.base import Serializer
 from aea.protocols.oef.message import OEFMessage
 from aea.protocols.oef.models import Description, Query
