+++ examples/gym_skill/helpers.py
@@ -26,7 +26,7 @@ from typing import Any, Tuple, cast
 
 from aea.mail.base import Envelope
 from aea.skills.base import SkillContext
-from aea.protocols.base.message import Message
+from aea.protocols.base import Message
 from aea.protocols.gym.message import GymMessage
 from aea.protocols.gym.serialization import GymSerializer
